# Conflicts:
#	app/src/main/java/com/example/citymap/MainActivity.kt

+++ app/src/main/java/com/example/citymap/IWeatherRepository.kt
@@ -4,6 +4,6 @@ interface IWeatherRepository {
 
     suspend fun getDataWeather()
 
-    suspend fun sendDataWeather(latitude: Float, longitude: Float)
+    suspend fun sendDataWeather(latitude: Double, longitude: Double)
 
 }
